pyproject.toml
[tool.poetry]
name = "prsb-site"
version = "0.1.0"
description = "website for the Phinney Ridge Steel Band members"
authors = ["Jim Andress <jim.andress@gmail.com>"]
readme = "README.md"
packages = [{include = "*", from = "prsb"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
psycopg = {version = "^3.1.18", extras = ["binary"]}
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
pandas = "^2.2.2"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
